+++ .idea/.name
@@ -1,1 +1,1 @@
-MealSnap
+MealSnap_Prototype_2
