--- README.md
@@ -1,34 +1,0 @@
-# MealSnap Android App
-
-## Product Overview
-This app allows you catalogue your food pictures an organized way as well as share it with other people. 
-
-## Requirements
- - [Android Studio](https://developer.android.com/studio)
-
-## Release
-Release is an APK file. To install it on an android emulator, drag the APK file to the emulator. See section below for how to install an new android emulator if you are unfamiliar.
-
-## Getting Started
-
-### Opening Project on Android Studio
-1. Open Android Studio
-2. Select option for opening an existing project
-3. From open existing project prompt, navigate to the root of this repository and select open
-
-### Running App on Android Emulator
-
-#### Installing a new Emulator
-If this is the first time you're using Android Studio, you won't have an emulator already instead. To install a new emulator, following the following steps.
-1. From Android Studio, navigate to tools -> AVD Manager
-2. Select "Create Virtual Device" on bottom left side of panel
-3. Select your preferred device -> Select API Level 28 -> Click Finish
-
-#### Running App on an Emulator
-1. From the top toolbar of the IDE, right before the green run button, you can select your preferred emulator (optional)
-2. Select Run on the top toolbar
-
-## References
-- [Backend](https://github.com/CSCI-40500-77100-Spring-2021/project-10__backend)
-- [Kanban Board](https://github.com/orgs/CSCI-40500-77100-Spring-2021/projects/1)
-- [Mock UI Previews](https://projects.invisionapp.com/share/HJ10DN9S6MRN#/screens/446418641?browse): Each screen has the following prefixes for each screens `Common`, `Android`, and `iOS`. For this project, we'll be implementing the `Common` and `Android` prefixed mock ui's.
