+++ app/build.gradle
@@ -7,7 +7,7 @@ android {
     buildToolsVersion "30.0.3"
 
     defaultConfig {
-        applicationId "com.example.mealsnap_prototype_1"
+        applicationId "com.example.mealsnap_prototype_2"
         minSdkVersion 21
         targetSdkVersion 30
         versionCode 1
@@ -25,10 +25,6 @@ android {
     compileOptions {
         sourceCompatibility JavaVersion.VERSION_1_8
         targetCompatibility JavaVersion.VERSION_1_8
-
-        coreLibraryDesugaringEnabled true
-        sourceCompatibility JavaVersion.VERSION_1_8
-        targetCompatibility JavaVersion.VERSION_1_8
     }
 }
 
@@ -40,24 +36,4 @@ dependencies {
     testImplementation 'junit:junit:4.+'
     androidTestImplementation 'androidx.test.ext:junit:1.1.2'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.3.0'
-
-    // Amplify core dependency
-    implementation 'com.amplifyframework:core:1.17.0'
-
-    // Support for Java 8 features
-    coreLibraryDesugaring 'com.android.tools:desugar_jdk_libs:1.1.5'
-
-    implementation 'com.amplifyframework:aws-auth-cognito:1.17.0'
-
-    implementation("com.squareup.okhttp3:okhttp:4.9.0")
-    implementation 'com.google.code.gson:gson:2.8.6'
-
-    implementation "com.github.parse-community.Parse-SDK-Android:parse:1.18.5"
-
-    implementation 'com.github.bumptech.glide:glide:4.11.0'
-    // Glide v4 uses this new annotation processor -- see https://bumptech.github.io/glide/doc/generatedapi.html
-    annotationProcessor 'com.github.bumptech.glide:compiler:4.11.0'
-
-    implementation 'androidx.swiperefreshlayout:swiperefreshlayout:1.0.0'
-
 }
