+++ app/src/androidTest/java/com/example/mealsnap_prototype_2/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package com.example.mealsnap_prototype_1;
+package com.example.mealsnap_prototype_2;
 
 import android.content.Context;
 
@@ -21,6 +21,6 @@ public class ExampleInstrumentedTest {
     public void useAppContext() {
         // Context of the app under test.
         Context appContext = InstrumentationRegistry.getInstrumentation().getTargetContext();
-        assertEquals("com.example.mealsnap_prototype_1", appContext.getPackageName());
+        assertEquals("com.example.mealsnap_prototype_2", appContext.getPackageName());
     }
 }
