+++ app/src/main/AndroidManifest.xml
@@ -1,44 +1,23 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.example.mealsnap_prototype_1">
-
-    <uses-permission android:name="android.permission.CAMERA" />
-    <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE" />
-    <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" />
-    <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
-    <uses-permission android:name="android.permission.INTERNET" />
+    package="com.example.mealsnap_prototype_2">
 
     <application
-        android:usesCleartextTraffic="true"
-        android:name=".ParseApplication"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.MealSnap_Prototype_1">
-        <activity android:name=".OnboardingActivity"/>
-        <activity android:name=".SignupActivity"/>
-        <activity android:name=".FeedActivity" />
-        <activity android:name=".LoginActivity" />
-        <activity android:name=".PostActivity" />
-        <activity android:name=".ConfigActivity">
+        android:theme="@style/Theme.MealSnap_Prototype_2">
+        <activity android:name=".OnboardingActivity"></activity>
+        <activity android:name=".ConfigActivity" />
+        <activity android:name=".MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-
-        <provider
-            android:name="androidx.core.content.FileProvider"
-            android:authorities="com.example.android.fileprovider"
-            android:exported="false"
-            android:grantUriPermissions="true">
-            <meta-data
-                android:name="android.support.FILE_PROVIDER_PATHS"
-                android:resource="@xml/file_paths" />
-        </provider>
     </application>
 
 </manifest>
