--- app/src/main/java/com/example/mealsnap_prototype_1/OnboardingActivity.java
@@ -1,60 +1,0 @@
-package com.example.mealsnap_prototype_1;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.util.Log;
-import android.view.View;
-import android.widget.Button;
-import android.widget.Toast;
-
-public class OnboardingActivity extends AppCompatActivity {
-    //Tag for debugging
-    public static final String TAG = "OnBoardingActivity";
-
-    Button loginButton;
-    Button signUpButton;
-
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_onboarding);
-
-        loginButton = findViewById(R.id.loginButton);
-        signUpButton = findViewById(R.id.signUpButton);
-
-        //on click to move to signup activity
-        signUpButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Log.i(TAG, "onClick Sign Up Button");
-                goToSignupActivity();
-            }
-        });
-
-        //on click move to login activity
-        loginButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Log.i(TAG, "onClick Sign Up Button");
-                goToLoginActivity();
-            }
-        });
-    }
-
-    //moves to LoginActivity
-    private void goToLoginActivity() {
-        Intent i = new Intent(this, LoginActivity.class);
-        startActivity(i);
-        finish();
-    }
-
-    //move to signUpActivity
-    private void goToSignupActivity() {
-        Intent i = new Intent(this, SignupActivity.class);
-        startActivity(i);
-        finish();
-    }
-}
