--- app/src/main/java/com/example/mealsnap_prototype_1/ParsePost.java
@@ -1,51 +1,0 @@
-////PARSE STUFF TO BE REMOVED
-package com.example.mealsnap_prototype_1;
-
-import com.parse.Parse;
-import com.parse.ParseClassName;
-import com.parse.ParseFile;
-import com.parse.ParseObject;
-import com.parse.ParseUser;
-
-@ParseClassName("Post") //string here is supposed to match with database name
-public class ParsePost extends ParseObject {
-    public static final String KEY_DESCRIPTION = "description";
-    public static final String KEY_IMAGE = "image";
-    public static final String KEY_TITLE = "title";
-    public static final String KEY_USER = "user";
-    public static final String KEY_CREATED_AT = "createdAt";
-
-
-    public String getDescription(){
-        return getString(KEY_DESCRIPTION);
-    }
-
-    public void setDescription(String description){
-        put(KEY_DESCRIPTION, description);
-    }
-
-    public ParseFile getImage(){
-        return getParseFile(KEY_IMAGE);
-    }
-
-    public void setImage(ParseFile parseFile){
-        put(KEY_IMAGE, parseFile);
-    }
-
-    public String getTitle(){
-        return getString(KEY_TITLE);
-    }
-
-    public void setTitle(String title){
-        put(KEY_TITLE, title);
-    }
-
-    public ParseUser getUser(){
-        return getParseUser(KEY_USER);
-    }
-
-    public void setUser(ParseUser user){
-        put(KEY_USER, user);
-    }
-}
-////PARSE STUFF TO BE REMOVED
