--- app/src/main/java/com/example/mealsnap_prototype_1/PostActivity.java
@@ -1,202 +1,0 @@
-package com.example.mealsnap_prototype_1;
-
-import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.core.app.ActivityCompat;
-import androidx.core.content.ContextCompat;
-import androidx.core.content.FileProvider;
-
-import android.Manifest;
-import android.app.Activity;
-import android.content.Intent;
-import android.content.pm.PackageManager;
-import android.net.Uri;
-import android.os.Bundle;
-import android.os.Environment;
-import android.provider.MediaStore;
-import android.util.Log;
-import android.view.View;
-import android.widget.Button;
-import android.widget.EditText;
-import android.widget.ImageView;
-import android.widget.Toast;
-
-import com.parse.FindCallback;
-import com.parse.ParseException;
-import com.parse.ParseFile;
-import com.parse.ParseQuery;
-import com.parse.ParseUser;
-import com.parse.SaveCallback;
-
-import java.io.File;
-import java.io.IOException;
-import java.text.SimpleDateFormat;
-import java.util.Date;
-import java.util.List;
-
-public class PostActivity extends AppCompatActivity {
-
-    public static final String TAG = "PostActivity";
-    public static final int CAMERA_PERMISSION_CODE = 101;
-    public static final int CAMERA_REQUEST_CODE = 102;
-
-    //declared xml stuff
-    private EditText etTitle;
-    private EditText etDescription;
-    private Button btnCaptureImage;
-    private ImageView ivPostImage;
-    private Button btnPost;
-    public File photoFile;
-    public  String currentPhotoPath;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_post);
-
-        //link xml stuff
-        etTitle = findViewById(R.id.etTitle);
-        etDescription = findViewById(R.id.etDescription);
-        btnCaptureImage = findViewById(R.id.btnCaptureImage);
-        ivPostImage = findViewById(R.id.ivPostImage);
-        btnPost = findViewById(R.id.btnPost);
-
-        btnCaptureImage.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                askCameraPermission();
-            }
-        });
-
-        ////PARSE STUFF TO BE REMOVED
-        btnPost.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                String description = etDescription.getText().toString();
-                String title = etTitle.getText().toString();
-                if(description.isEmpty()){
-                    Toast.makeText(PostActivity.this, "Description can't be empty", Toast.LENGTH_SHORT).show();
-                    return;
-                }
-                if(title.isEmpty()){
-                    Toast.makeText(PostActivity.this, "Title can't be empty", Toast.LENGTH_SHORT).show();
-                    return;
-                }
-                ParseUser currentUser = ParseUser.getCurrentUser();
-                savePost(title, description, currentUser);
-                //goToFeedActivity();
-
-            }
-        });
-        ////PARSE STUFF TO BE REMOVED
-    }
-
-    ////PARSE STUFF TO BE REMOVED
-    private void savePost(String title, String description, ParseUser currentUser) {
-        ParsePost post = new ParsePost();
-        post.setDescription(description);
-        post.setImage(new ParseFile(photoFile)); //super hardcoded, photofile is currently being overwritten
-        post.setTitle(title);
-        post.setUser(currentUser);
-        post.saveInBackground(new SaveCallback() {
-            @Override
-            public void done(ParseException e) {
-                if(e != null){
-                    Log.e(TAG, "Error while saving", e);
-                    Toast.makeText(PostActivity.this, "Failed to save.", Toast.LENGTH_SHORT).show();
-                }
-                Log.i(TAG, "Post was saved successfully!");
-                etDescription.setText("");
-                etTitle.setText("");
-                ivPostImage.setImageResource(0);
-                finish();
-            }
-        });
-    }
-    ////PARSE STUFF TO BE REMOVED
-
-
-    //ask for camera permission
-    private void askCameraPermission() {
-        if(ContextCompat.checkSelfPermission(this, Manifest.permission.CAMERA) != PackageManager.PERMISSION_GRANTED){
-            //If we dont have permission request it
-            ActivityCompat.requestPermissions(this, new String[] {Manifest.permission.CAMERA}, CAMERA_PERMISSION_CODE);
-        }
-        else{
-            dispatchTakePictureIntent();
-        }
-    }
-
-    @Override
-    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
-        //super.onRequestPermissionsResult(requestCode, permissions, grantResults);
-        if(requestCode == CAMERA_PERMISSION_CODE){
-            if(grantResults.length > 0 && grantResults[0] == PackageManager.PERMISSION_GRANTED){
-                dispatchTakePictureIntent();
-            }else{
-                Toast.makeText(PostActivity.this,"Camera Permission was Denied", Toast.LENGTH_SHORT).show();
-            }
-        }
-    }
-
-    //@SuppressLint("MissingSuperCall") worked as well
-    @Override
-    protected void onActivityResult(int requestCode, int resultCode, @Nullable Intent data) {
-        super.onActivityResult(requestCode, resultCode, data);
-        if(requestCode == CAMERA_REQUEST_CODE) {
-            if(resultCode == Activity.RESULT_OK){
-                File f = new File(currentPhotoPath);
-                ivPostImage.setImageURI(Uri.fromFile(f));
-                Log.d(TAG, "Absolute Url of image is: " + Uri.fromFile(f));
-            }
-        }
-    }
-
-    //taken from android documentation https://developer.android.com/training/camera/photobasics#java
-    private File createImageFile() throws IOException {
-        // Create an image file name
-        String timeStamp = new SimpleDateFormat("yyyyMMdd_HHmmss").format(new Date());
-        String imageFileName = "JPEG_" + timeStamp + "_";
-        File storageDir = getExternalFilesDir(Environment.DIRECTORY_PICTURES); //stored here not in gallery, gallery method has some issues with this android lollipop version
-        File image = File.createTempFile(
-                imageFileName,  /* prefix */
-                ".jpg",         /* suffix */
-                storageDir      /* directory */
-        );
-        // Save a file: path for use with ACTION_VIEW intents
-        currentPhotoPath = image.getAbsolutePath();
-        photoFile = image;
-        return image;
-    }
-
-    //taken from android documentation https://developer.android.com/training/camera/photobasics#java
-    private void dispatchTakePictureIntent() {
-        Intent takePictureIntent = new Intent(MediaStore.ACTION_IMAGE_CAPTURE);
-        // Ensure that there's a camera activity to handle the intent
-        if (takePictureIntent.resolveActivity(getPackageManager()) != null) {
-            // Create the File where the photo should go
-            File photoFile = null;
-            try {
-                photoFile = createImageFile();
-            } catch (IOException ex) {
-                // Error occurred while creating the File
-            }
-            // Continue only if the File was successfully created
-            if (photoFile != null) {
-                Uri photoURI = FileProvider.getUriForFile(this,
-                        "com.example.android.fileprovider",
-                        photoFile);
-                takePictureIntent.putExtra(MediaStore.EXTRA_OUTPUT, photoURI);
-                startActivityForResult(takePictureIntent, CAMERA_REQUEST_CODE);
-            }
-        }
-    }
-
-    //moves to FeedActivity
-    private void goToFeedActivity() {
-        Intent i = new Intent(this, FeedActivity.class);
-        startActivity(i);
-        finish();
-    }
-}
