+++ app/src/main/java/com/example/mealsnap_prototype_2/OnboardingActivity.java
@@ -1,0 +1,14 @@
+package com.example.mealsnap_prototype_2;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.os.Bundle;
+
+public class OnboardingActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_onboarding);
+    }
+}
