--- app/src/main/java/services/authhandler/APIServiceResponseEvent.java
@@ -1,11 +1,0 @@
-package services.authhandler;
-
-import java.io.IOException;
-
-import okhttp3.RequestBody;
-import okhttp3.ResponseBody;
-
-public interface APIServiceResponseEvent{
-    void onSuccess(ResponseBody requestBody);
-    void onFailure(IOException ioException);
-}
