--- app/src/main/java/services/authhandler/AuthHandler.java
@@ -1,17 +1,0 @@
-package services.authhandler;
-
-
-import android.util.Log;
-
-public class AuthHandler {
-    /*
-    public static void signIn(String username, String password){
-        com.amplifyframework.core.Amplify.Auth.signIn(
-                "defaultuser", //username
-                "P@5sword", //password
-                result -> Log.i("AuthQuickstart", result.isSignInComplete() ? "Sign in succeeded" : "Sign in not complete"),
-                error -> Log.e("AuthQuickstart", error.toString())
-        );
-        //handle sign in move to feed page
-    }*/
-}
