--- app/src/main/res/drawable-anydpi/ic_action_name.xml
@@ -1,17 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="24dp"
-    android:height="24dp"
-    android:viewportWidth="24"
-    android:viewportHeight="24"
-    android:tint="#333333"
-    android:alpha="0.6"
-    android:autoMirrored="true">
-  <group android:scaleX="2"
-      android:scaleY="2"
-      android:translateX="-12"
-      android:translateY="-12">
-    <path
-        android:fillColor="@android:color/white"
-        android:pathData="M15.41,16.59L10.83,12l4.58,-4.59L14,6l-6,6 6,6 1.41,-1.41z"/>
-  </group>
-</vector>
