--- app/src/main/res/layout/item_parse_post.xml
@@ -1,46 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:padding="16dp"
-    android:layout_height="wrap_content">
-
-    <ImageView
-        android:id="@+id/ppImage"
-        android:layout_width="350dp"
-        android:layout_height="260dp"
-        android:layout_centerHorizontal="true"
-        tools:srcCompat="@tools:sample/avatars" />
-
-    <TextView
-        android:id="@+id/ppTitle"
-        style="@style/TextAppearance.AppCompat.Title"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_below="@+id/ppImage"
-        android:layout_marginStart="15dp"
-        android:textSize="30sp"
-        tools:text="sample" />
-
-    <TextView
-        android:id="@+id/ppUser"
-        style="@style/TextAppearance.AppCompat.Subhead"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_below="@+id/ppTitle"
-        android:layout_marginStart="15dp"
-        android:textSize="20sp"
-        tools:text="sample" />
-
-    <TextView
-        android:id="@+id/ppDesc"
-        style="@style/TextAppearance.AppCompat.Subhead"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_below="@+id/ppUser"
-        android:layout_marginStart="15dp"
-        android:textSize="16sp"
-        tools:text="sample" />
-
-</RelativeLayout>
