--- app/src/main/res/layout/post_view.xml
@@ -1,36 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:padding="16dp"
-    android:layout_height="match_parent">
-
-    <ImageView
-        android:id="@+id/pImage"
-        android:layout_width="350dp"
-        android:layout_height="260dp"
-        android:layout_centerHorizontal="true"
-        tools:srcCompat="@tools:sample/avatars" />
-
-    <TextView
-        android:id="@+id/pTitle"
-        style="@style/TextAppearance.AppCompat.Title"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_below="@+id/pImage"
-        android:layout_marginStart="15dp"
-        android:textSize="30sp"
-        tools:text="sample" />
-
-    <TextView
-        android:id="@+id/pDesc"
-        style="@style/TextAppearance.AppCompat.Subhead"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_below="@+id/pTitle"
-        android:layout_marginStart="15dp"
-        android:textSize="18sp"
-        tools:text="sample" />
-
-</RelativeLayout>
