--- app/src/main/res/raw/amplifyconfiguration.json
@@ -1,23 +1,0 @@
-{
-  "auth": {
-    "plugins": {
-      "awsCognitoAuthPlugin": {
-        "IdentityManager": {
-          "Default": {}
-        },
-        "CognitoUserPool": {
-          "Default": {
-            "PoolId": "us-east-1_ANFViz2yG",
-            "AppClientId": "32ldvognhrlumro111169lbt9h",
-            "Region": "us-east-1"
-          }
-        },
-        "Auth": {
-          "Default": {
-            "authenticationFlowType": "USER_SRP_AUTH"
-          }
-        }
-      }
-    }
-  }
-}
