+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">MealSnap_Prototype_1</string>
+    <string name="app_name">MealSnap_Prototype_2</string>
 </resources>
