--- app/src/main/res/xml/file_paths.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<paths>
-    <external-path
-        name="external"
-        path="."/>
-    <external-files-path
-        name="external_files"
-        path="."/>
-    <cache-path
-        name="cache"
-        path="."/>
-    <external-cache-path
-        name="external_cache"
-        path="."/>
-    <files-path
-        name="files"
-        path="."/>
-</paths>
