+++ build.gradle
@@ -16,8 +16,6 @@ allprojects {
     repositories {
         google()
         jcenter()
-        maven { url "https://jitpack.io" }
-
     }
 }
