+++ settings.gradle
@@ -1,2 +1,2 @@
 include ':app'
-rootProject.name = "MealSnap"
+rootProject.name = "MealSnap_Prototype_2"
